Refactor: relocate `envparse` module to `internal/envparse` for shared use across server and agent, removing duplication and consolidating validation logic.
