+++ .github/workflows/go-lint.yml
@@ -4,14 +4,16 @@ on:
   push:
     branches: [main]
     paths:
-      - "agent/**"
-      - "server/**"
+      - "**/*.go"
+      - "go.mod"
+      - "go.sum"
       - ".golangci.yml"
       - ".github/workflows/go-lint.yml"
   pull_request:
     paths:
-      - "agent/**"
-      - "server/**"
+      - "**/*.go"
+      - "go.mod"
+      - "go.sum"
       - ".golangci.yml"
       - ".github/workflows/go-lint.yml"
   workflow_dispatch:
@@ -26,28 +28,21 @@ concurrency:
 jobs:
   golangci-lint:
     name: golangci-lint
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        include:
-          - module: agent
-            os: macos-latest
-          - module: server
-            os: ubuntu-latest
+    # macos-latest so the agent's CGo build tags (macOS-only Endpoint Security /
+    # XPC bridge code) compile; ubuntu would fail on the CGo headers.
+    runs-on: macos-latest
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
         with:
           persist-credentials: false
 
       - name: Create UI dist placeholder for embed
-        if: matrix.module == 'server'
         run: mkdir -p server/ui/dist && touch server/ui/dist/.gitkeep
 
       - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6.4.0
         with:
-          go-version-file: ${{ matrix.module }}/go.mod
+          go-version-file: go.mod
 
       - uses: golangci/golangci-lint-action@1481404843c368bc19ca9406f87d6e0fc97bdcfd # v7.0.0
         with:
           version: v2.11
-          working-directory: ${{ matrix.module }}
