--- envparse/go.mod
@@ -1,3 +1,0 @@
-module github.com/fleetdm/edr/envparse
-
-go 1.26
