+++ server/config/config.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/fleetdm/edr/envparse"
+	"github.com/fleetdm/edr/internal/envparse"
 )
 
 // Config is the resolved server configuration.
